loosing sight on what im doing
