+++ nodejs-proxy/src/index.js
@@ -1,48 +1,23 @@
 import { Server } from "socket.io";
 import WebSocket from "ws";
+import { socketIoHandler } from "./ws-handler/socket-io/index.js";
+import { goSocketHandler } from "./ws-handler/go-server/index.js";
 
+const HOST = "0.0.0.0";
+const PORT = 3001;
 const clients = {};
-const io = new Server(3001, {
-    cors: { origin: "*" },
-});
+const main = async () => {
+    try {
+        const io = new Server(PORT, {
+            cors: { origin: "*" },
+        });
+        const goSocket = new WebSocket("ws://localhost:9000/ws"); // Connect to Go
 
-const goSocket = new WebSocket("ws://localhost:4000/ws"); // Connect to Go
-
-// --- Relay messages from GO to client(s) ---
-goSocket.on("message", (data) => {
-    const { clientId, event, payload } = JSON.parse(data);
-    const client = clients[clientId];
-    if (client) {
-        client.emit(event, payload);
+        console.log(`Server running on: http://${HOST}:${PORT}/`);
+        goSocketHandler(goSocket, clients);
+        socketIoHandler(goSocket, io, clients);
+    } catch (err) {
+        console.error("Error on node bridge: ", err);
     }
-});
-
-io.on("connection", (socket) => {
-    console.log("Client connected:", socket.id);
-    clients[socket.id] = socket;
-
-    socket.onAny((event, payload) => {
-        console.log("Event: ", event);
-        console.log("Payload: ", payload);
-        // Forward to Go
-        const ctx = {
-            clientId: socket.id,
-            event,
-            payload,
-        };
-        console.log("Sending message to Go socket: ", ctx);
-        const msg = JSON.stringify(ctx);
-        goSocket.send(msg);
-    });
-
-    socket.on("disconnect", () => {
-        delete clients[socket.id];
-        const ctx = {
-            clientId: socket.id,
-            event: "disconnect",
-        };
-        console.log("Disconneted socket: ", ctx);
-        const msg = JSON.stringify(ctx);
-        goSocket.send(msg);
-    });
-});
+};
+main();
